# More efficient JSON decoders and encoders for records